requirements.txt
python-telegram-bot==22.5
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn[standard]==0.34.0
